fix(wallet): use BigDecimal for monetary amounts in DTOs
Currency amounts should never round-trip through Double. Refactor
CreateTransactionRequest.amount, CreateRecurringRequest.amount, and
SettleDebtRequest.amount from Double to BigDecimal. Drop BigDecimal.valueOf(double)
wrappers in TransactionService, RecurringService, DebtGroupService — Jackson
deserializes JSON numbers directly into BigDecimal using their text representation,
which is exact.

Defense-in-depth against the kind of off-by-one precision bug seen in old
transaction data (e.g. 20000 stored as 19999, 100000000 stored as 99999998).

+++ src/main/java/dev/chinh/portfolio/apps/wallet/RecurringService.java
@@ -32,7 +32,7 @@ public class RecurringService {
         rule.setUserId(userId);
         rule.setWalletId(req.walletId());
         rule.setCategoryId(req.categoryId());
-        rule.setAmount(BigDecimal.valueOf(req.amount()));
+        rule.setAmount(req.amount());
         rule.setType(req.type());
         rule.setFrequency(req.frequency());
         rule.setStartDate(LocalDate.parse(req.startDate()));
@@ -54,7 +54,7 @@ public class RecurringService {
         }
         rule.setWalletId(req.walletId());
         rule.setCategoryId(req.categoryId());
-        rule.setAmount(BigDecimal.valueOf(req.amount()));
+        rule.setAmount(req.amount());
         rule.setType(req.type());
         rule.setFrequency(req.frequency());
         rule.setStartDate(LocalDate.parse(req.startDate()));
