+++ src/main/java/dev/chinh/portfolio/apps/wallet/dto/SettleDebtRequest.java
@@ -1,9 +1,10 @@
 package dev.chinh.portfolio.apps.wallet.dto;
 
 import jakarta.validation.constraints.*;
+import java.math.BigDecimal;
 
 public record SettleDebtRequest(
-    @NotNull @DecimalMin("0.01") Double amount,
+    @NotNull @DecimalMin("0.01") BigDecimal amount,
     @NotNull Long walletId,
     String note
 ) {}
